put test data to one directory
