+++ src/components/movie-card/movie-card.e2e.test.js
@@ -2,22 +2,8 @@ import React from "react";
 import Enzyme, {shallow} from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 import MovieCard from './movie-card.jsx';
+import {MOVIE} from "../../data-for-tests/data-for-tests";
 
-const mocksData = {
-  title: `Fantastic Beasts: The Crimes of Grindelwald`,
-  genre: `Drama`,
-  release: 2017,
-  img: `img/fantastic-beasts-the-crimes-of-grindelwald.jpg`,
-  bg: `img/bg-the-grand-budapest-hotel.jpg`,
-  poster: `img/the-grand-budapest-hotel-poster.jpg`,
-  description: `In the 1930s, the Grand Budapest Hotel is a popular European ski resort, presided over by concierge Gustave H. (Ralph Fiennes). Zero, a junior lobby boy, becomes Gustave's friend and protege.`,
-  director: `Wes Andreson`,
-  starring: [`Bill Murray`, `Edward Norton`, `Jude Law`, `Willem Dafoe`],
-  ratingScore: 8.9,
-  ratingCount: 240,
-  id: 1,
-  src: `https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4`,
-};
 const onMovieCardHover = jest.fn((args) => args);
 
 Enzyme.configure({
@@ -27,7 +13,7 @@ Enzyme.configure({
 it(`Should get MovieCard hovered`, () => {
   const component = shallow(
       <MovieCard
-        movie = {mocksData}
+        movie = {MOVIE}
         onMovieClick = {() => {}}
         onMovieCardHover = {onMovieCardHover}
       />
@@ -36,7 +22,5 @@ it(`Should get MovieCard hovered`, () => {
   const card = component.find(`.small-movie-card`);
   card.simulate(`mouseover`);
   expect(onMovieCardHover).toHaveBeenCalledTimes(1);
-  // onMovieCardHover.mockImplementation(() => mocksData);
-  // expect(onMovieCardHover()).toBe(mocksData);
-  expect(onMovieCardHover).toHaveBeenCalledWith(mocksData);
+  expect(onMovieCardHover).toHaveBeenCalledWith(MOVIE);
 });
