+++ src/components/movie-list/movie-list.test.js
@@ -1,131 +1,7 @@
 import React from "react";
 import renderer from "react-test-renderer";
 import MovieList from "./movie-list.jsx";
-
-export const MOVIES = [
-  {
-    title: `Fantastic Beasts: The Crimes of Grindelwald`,
-    genre: `Drama`,
-    release: 2017,
-    img: `img/fantastic-beasts-the-crimes-of-grindelwald.jpg`,
-    bg: `img/bg-the-grand-budapest-hotel.jpg`,
-    poster: `img/the-grand-budapest-hotel-poster.jpg`,
-    description: `In the 1930s, the Grand Budapest Hotel is a popular European ski resort, presided over by concierge Gustave H. (Ralph Fiennes). Zero, a junior lobby boy, becomes Gustave's friend and protege.`,
-    director: `Wes Andreson`,
-    starring: [`Bill Murray`, `Edward Norton`, `Jude Law`, `Willem Dafoe`],
-    ratingScore: 8.9,
-    ratingCount: 240,
-    id: 1,
-    src: `https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4`,
-  },
-  {
-    title: `Bohemian Rhapsody`,
-    genre: `Drama`,
-    release: 2017,
-    img: `img/fantastic-beasts-the-crimes-of-grindelwald.jpg`,
-    bg: `img/bg-the-grand-budapest-hotel.jpg`,
-    poster: `img/the-grand-budapest-hotel-poster.jpg`,
-    description: `In the 1930s, the Grand Budapest Hotel is a popular European ski resort, presided over by concierge Gustave H. (Ralph Fiennes). Zero, a junior lobby boy, becomes Gustave's friend and protege.`,
-    director: `Wes Andreson`,
-    starring: [`Bill Murray`, `Edward Norton`, `Jude Law`, `Willem Dafoe`],
-    ratingScore: 8.9,
-    ratingCount: 240,
-    id: 2,
-    src: `https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4`,
-  },
-  {
-    title: `Macbeth`,
-    genre: `Drama`,
-    release: 2017,
-    img: `img/fantastic-beasts-the-crimes-of-grindelwald.jpg`,
-    bg: `img/bg-the-grand-budapest-hotel.jpg`,
-    poster: `img/the-grand-budapest-hotel-poster.jpg`,
-    description: `In the 1930s, the Grand Budapest Hotel is a popular European ski resort, presided over by concierge Gustave H. (Ralph Fiennes). Zero, a junior lobby boy, becomes Gustave's friend and protege.`,
-    director: `Wes Andreson`,
-    starring: [`Bill Murray`, `Edward Norton`, `Jude Law`, `Willem Dafoe`],
-    ratingScore: 8.9,
-    ratingCount: 240,
-    id: 3,
-    src: `https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4`,
-  },
-  {
-    title: `Aviator`,
-    genre: `Drama`,
-    release: 2017,
-    img: `img/fantastic-beasts-the-crimes-of-grindelwald.jpg`,
-    bg: `img/bg-the-grand-budapest-hotel.jpg`,
-    poster: `img/the-grand-budapest-hotel-poster.jpg`,
-    description: `In the 1930s, the Grand Budapest Hotel is a popular European ski resort, presided over by concierge Gustave H. (Ralph Fiennes). Zero, a junior lobby boy, becomes Gustave's friend and protege.`,
-    director: `Wes Andreson`,
-    starring: [`Bill Murray`, `Edward Norton`, `Jude Law`, `Willem Dafoe`],
-    ratingScore: 8.9,
-    ratingCount: 240,
-    id: 4,
-    src: `https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4`,
-  },
-  {
-    title: `We need to talk about Kevin`,
-    genre: `Drama`,
-    release: 2017,
-    img: `img/fantastic-beasts-the-crimes-of-grindelwald.jpg`,
-    bg: `img/bg-the-grand-budapest-hotel.jpg`,
-    poster: `img/the-grand-budapest-hotel-poster.jpg`,
-    description: `In the 1930s, the Grand Budapest Hotel is a popular European ski resort, presided over by concierge Gustave H. (Ralph Fiennes). Zero, a junior lobby boy, becomes Gustave's friend and protege.`,
-    director: `Wes Andreson`,
-    starring: [`Bill Murray`, `Edward Norton`, `Jude Law`, `Willem Dafoe`],
-    ratingScore: 8.9,
-    ratingCount: 240,
-    id: 5,
-    src: `https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4`,
-  },
-  {
-    title: `What We Do in the Shadows`,
-    genre: `Drama`,
-    release: 2017,
-    img: `img/fantastic-beasts-the-crimes-of-grindelwald.jpg`,
-    bg: `img/bg-the-grand-budapest-hotel.jpg`,
-    poster: `img/the-grand-budapest-hotel-poster.jpg`,
-    description: `In the 1930s, the Grand Budapest Hotel is a popular European ski resort, presided over by concierge Gustave H. (Ralph Fiennes). Zero, a junior lobby boy, becomes Gustave's friend and protege.`,
-    director: `Wes Andreson`,
-    starring: [`Bill Murray`, `Edward Norton`, `Jude Law`, `Willem Dafoe`],
-    ratingScore: 8.9,
-    ratingCount: 240,
-    id: 6,
-    src: `https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4`,
-  },
-  {
-    title: `Pulp Fiction`,
-    genre: `Drama`,
-    release: 2017,
-    img: `img/fantastic-beasts-the-crimes-of-grindelwald.jpg`,
-    bg: `img/bg-the-grand-budapest-hotel.jpg`,
-    poster: `img/the-grand-budapest-hotel-poster.jpg`,
-    description: `In the 1930s, the Grand Budapest Hotel is a popular European ski resort, presided over by concierge Gustave H. (Ralph Fiennes). Zero, a junior lobby boy, becomes Gustave's friend and protege.`,
-    director: `Wes Andreson`,
-    starring: [`Bill Murray`, `Edward Norton`, `Jude Law`, `Willem Dafoe`],
-    ratingScore: 8.9,
-    ratingCount: 240,
-    id: 7,
-    src: `https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4`,
-  },
-  {
-    title: `Snatch`,
-    genre: `Drama`,
-    release: 2017,
-    img: `img/fantastic-beasts-the-crimes-of-grindelwald.jpg`,
-    bg: `img/bg-the-grand-budapest-hotel.jpg`,
-    poster: `img/the-grand-budapest-hotel-poster.jpg`,
-    description: `In the 1930s, the Grand Budapest Hotel is a popular European ski resort, presided over by concierge Gustave H. (Ralph Fiennes). Zero, a junior lobby boy, becomes Gustave's friend and protege.`,
-    director: `Wes Andreson`,
-    starring: [`Bill Murray`, `Edward Norton`, `Jude Law`, `Willem Dafoe`],
-    ratingScore: 8.9,
-    ratingCount: 240,
-    id: 8,
-    src: `https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4`,
-  }
-];
-
-// const onMovieClick = () => {};
+import {MOVIES} from "../../data-for-tests/data-for-tests.js";
 
 it(`Should MovieList render Correctly`, () => {
   const tree = renderer
