+++ src/components/video-player/video-player.test.js
@@ -1,19 +1,15 @@
 import React from "react";
 import renderer from "react-test-renderer";
 import VideoPlayer from "./video-player.jsx";
-
-const video = {
-  poster: `img/the-grand-budapest-hotel-poster.jpg`,
-  src: `https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4`,
-};
+import {VIDEO} from "../../data-for-tests/data-for-tests.js";
 
 describe(`VideoPlayer`, () => {
   it(`VideoPlayer should render correctly`, () => {
     const tree = renderer
       .create(<VideoPlayer
         isPlaying={false}
-        src={video.src}
-        poster={video.poster}
+        src={VIDEO.src}
+        poster={VIDEO.poster}
       />, {
         createNodeMock: () => {
           return {};
